Cleanup to make autofix-in-seer more explicit
