+++ src/sentry/seer/autofix/autofix_agent.py
@@ -4,7 +4,6 @@ import logging
 import re
 from collections.abc import Sequence
 from dataclasses import dataclass
-from enum import StrEnum
 from typing import TYPE_CHECKING, Any, Literal, cast
 
 import sentry_sdk
@@ -48,6 +47,7 @@ from sentry.seer.autofix.prompts import (
     root_cause_prompt,
     solution_prompt,
 )
+from sentry.seer.autofix.steps import AutofixStep
 from sentry.seer.autofix.types import AutofixHandoffResponse
 from sentry.seer.autofix.utils import (
     AutofixStoppingPoint,
@@ -87,34 +87,6 @@ class NoSeerQuotaException(Exception):
 
 class PrIterationNoPullRequestException(Exception):
     pass
-
-
-class AutofixStep(StrEnum):
-    """Available autofix steps."""
-
-    ROOT_CAUSE = "root_cause"
-    SOLUTION = "solution"
-    CODE_CHANGES = "code_changes"
-    PR_ITERATION = "pr_iteration"
-
-    @staticmethod
-    def from_autofix_stopping_point(
-        autofix_stopping_point: AutofixStoppingPoint,
-    ) -> AutofixStep:
-        match autofix_stopping_point:
-            case AutofixStoppingPoint.ROOT_CAUSE:
-                return AutofixStep.ROOT_CAUSE
-            case AutofixStoppingPoint.SOLUTION:
-                return AutofixStep.SOLUTION
-            case AutofixStoppingPoint.CODE_CHANGES:
-                return AutofixStep.CODE_CHANGES
-            case AutofixStoppingPoint.OPEN_PR:
-                # This depends on the last step being
-                # code changes and we should look for
-                # the PR elsewhere in the agent results
-                return AutofixStep.CODE_CHANGES
-            case _:
-                raise ValueError(f"Unsupported AutofixStoppingPoint: {autofix_stopping_point}")
 
 
 class StepConfig:
@@ -492,6 +464,54 @@ def _build_base_shas_metadata(group: Group, referrer: AutofixReferrer) -> str | 
     return json.dumps(base_shas)
 
 
+def trigger_autofix_rca_in_seer(
+    group: Group,
+    *,
+    referrer: AutofixReferrer,
+    user_context: str | None,
+    stopping_point: AutofixStoppingPoint | None,
+    allow_free_cohort: bool,
+    user: User | RpcUser | AnonymousUser | None,
+    enable_bash_tools: bool,
+) -> SeerRun:
+    """Start the RCA-in-Seer feature run and emit the legacy start events."""
+    # Local import avoids a circular import (dispatch imports this module).
+    from sentry.seer.autofix_rca.dispatch import trigger_autofix_rca_feature
+
+    feature_run = trigger_autofix_rca_feature(
+        group,
+        referrer=referrer,
+        user_context=user_context,
+        stopping_point=stopping_point,
+        allow_free_cohort=allow_free_cohort,
+        user=user,
+        enable_bash_tools=enable_bash_tools,
+    )
+    feature_run_id = feature_run.seer_run_state_id
+    if feature_run_id is None:
+        # flush=True populates this on success; guard defensively.
+        raise SeerApiError("autofix_rca feature run has no run id", 500)
+
+    logger.info(
+        "autofix.trigger.routed_to_rca_feature",
+        extra={
+            "group_id": group.id,
+            "organization_id": group.organization.id,
+            "run_id": feature_run_id,
+            "referrer": referrer.value,
+        },
+    )
+
+    _handle_step_started_events(
+        group,
+        AutofixStep.ROOT_CAUSE,
+        feature_run_id,
+        str(feature_run.uuid),
+        referrer,
+    )
+    return feature_run
+
+
 def trigger_autofix_agent(
     group: Group,
     step: AutofixStep,
@@ -544,10 +564,7 @@ def trigger_autofix_agent(
         "organizations:autofix-rca-in-seer", group.organization, actor=user
     )
     if step == AutofixStep.ROOT_CAUSE and run_id is None and use_seer_rca_feature:
-        # Local import avoids a circular import (dispatch imports this module).
-        from sentry.seer.autofix_rca.dispatch import trigger_autofix_rca_feature
-
-        feature_run = trigger_autofix_rca_feature(
+        return trigger_autofix_rca_in_seer(
             group,
             referrer=referrer,
             user_context=user_context,
@@ -556,29 +573,6 @@ def trigger_autofix_agent(
             user=user,
             enable_bash_tools=enable_bash_tools,
         )
-        feature_run_id = feature_run.seer_run_state_id
-        if feature_run_id is None:
-            # flush=True populates this on success; guard defensively.
-            raise SeerApiError("autofix_rca feature run has no run id", 500)
-
-        logger.info(
-            "autofix.trigger.routed_to_rca_feature",
-            extra={
-                "group_id": group.id,
-                "organization_id": group.organization.id,
-                "run_id": feature_run_id,
-                "referrer": referrer.value,
-            },
-        )
-
-        _handle_step_started_events(
-            group,
-            AutofixStep.ROOT_CAUSE,
-            feature_run_id,
-            str(feature_run.uuid),
-            referrer,
-        )
-        return feature_run
 
     config = STEP_CONFIGS[step]
