+++ src/sentry/seer/autofix/issue_summary.py
@@ -23,7 +23,6 @@ from sentry.models.group import Group
 from sentry.net.http import connection_from_url
 from sentry.seer.autofix.autofix import get_trace_tree_for_event
 from sentry.seer.autofix.autofix_agent import (
-    AutofixStep,
     NoSeerQuotaException,
     trigger_autofix_agent,
 )
@@ -33,6 +32,7 @@ from sentry.seer.autofix.constants import (
     FixabilityScoreThresholds,
     SeerAutomationSource,
 )
+from sentry.seer.autofix.steps import AutofixStep
 from sentry.seer.autofix.utils import (
     AutofixStoppingPoint,
     is_seer_autotriggered_autofix_rate_limited,
