+++ tests/sentry/seer/autofix/test_autofix_agent.py
@@ -14,7 +14,6 @@ from sentry.seer.agent.client_models import (
 )
 from sentry.seer.autofix.autofix_agent import (
     STEP_CONFIGS,
-    AutofixStep,
     NoSeerQuotaException,
     PrIterationNoPullRequestException,
     _build_base_shas_metadata,
@@ -29,6 +28,7 @@ from sentry.seer.autofix.autofix_agent import (
 )
 from sentry.seer.autofix.commit_author import SeerCommitAuthor
 from sentry.seer.autofix.constants import AutofixReferrer
+from sentry.seer.autofix.steps import AutofixStep
 from sentry.seer.autofix.utils import AutofixStoppingPoint
 from sentry.seer.models import SeerPermissionError
 from sentry.sentry_apps.utils.webhooks import SeerActionType
